Use AppNavigationHeader in app
