+++ desktop/packages/mullvad-vpn/src/renderer/components/ApiAccessMethods.tsx
@@ -14,6 +14,7 @@ import { generateRoutePath } from '../lib/routeHelpers';
 import { RoutePath } from '../lib/routes';
 import { useBoolean } from '../lib/utility-hooks';
 import { useSelector } from '../redux/store';
+import { AppNavigationHeader } from './';
 import * as Cell from './cell';
 import {
   ContextMenu,
@@ -26,13 +27,7 @@ import InfoButton from './InfoButton';
 import { BackAction } from './KeyboardNavigation';
 import { Layout, SettingsContainer, SettingsContent, SettingsNavigationScrollbars } from './Layout';
 import { ModalAlert, ModalAlertType } from './Modal';
-import {
-  NavigationBar,
-  NavigationContainer,
-  NavigationInfoButton,
-  NavigationItems,
-  TitleBarItem,
-} from './NavigationBar';
+import { NavigationContainer } from './NavigationContainer';
 import SettingsHeader, { HeaderSubTitle, HeaderTitle } from './SettingsHeader';
 import { SmallButton, SmallButtonColor } from './SmallButton';
 
@@ -42,7 +37,7 @@ const StyledContextMenuButton = styled(Cell.Icon)({
   marginRight: Spacings.spacing3,
 });
 
-const StyledMethodInfoButton = styled(InfoButton)({
+const StyledMethodInfoButton = styled(InfoButton).attrs({ size: 'small' })({
   marginRight: Spacings.spacing4,
 });
 
@@ -82,32 +77,28 @@ export default function ApiAccessMethods() {
       <Layout>
         <SettingsContainer>
           <NavigationContainer>
-            <NavigationBar>
-              <NavigationItems>
-                <TitleBarItem>
-                  {
-                    // TRANSLATORS: Title label in navigation bar
-                    messages.pgettext('navigation-bar', 'API access')
-                  }
-                </TitleBarItem>
-                <NavigationInfoButton
-                  message={[
-                    messages.pgettext(
-                      'api-access-methods-view',
-                      'The app needs to communicate with a Mullvad API server to log you in, fetch server lists, and other critical operations.',
-                    ),
-                    messages.pgettext(
-                      'api-access-methods-view',
-                      'On some networks, where various types of censorship are being used, the API servers might not be directly reachable.',
-                    ),
-                    messages.pgettext(
-                      'api-access-methods-view',
-                      'This feature allows you to circumvent that censorship by adding custom ways to access the API via proxies and similar methods.',
-                    ),
-                  ]}
-                />
-              </NavigationItems>
-            </NavigationBar>
+            <AppNavigationHeader
+              title={
+                // TRANSLATORS: Title label in navigation bar
+                messages.pgettext('navigation-bar', 'API access')
+              }>
+              <AppNavigationHeader.InfoButton
+                message={[
+                  messages.pgettext(
+                    'api-access-methods-view',
+                    'The app needs to communicate with a Mullvad API server to log you in, fetch server lists, and other critical operations.',
+                  ),
+                  messages.pgettext(
+                    'api-access-methods-view',
+                    'On some networks, where various types of censorship are being used, the API servers might not be directly reachable.',
+                  ),
+                  messages.pgettext(
+                    'api-access-methods-view',
+                    'This feature allows you to circumvent that censorship by adding custom ways to access the API via proxies and similar methods.',
+                  ),
+                ]}
+              />
+            </AppNavigationHeader>
 
             <SettingsNavigationScrollbars fillContainer>
               <SettingsContent>
