+++ desktop/packages/mullvad-vpn/src/renderer/components/index.ts
@@ -1,0 +1,1 @@
+export * from './app-navigation-header';
